requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
tensorflow-cpu==2.11.0
